+++ semana7/labeusers/src/App.js
@@ -1,85 +1,37 @@
 import "./App.css";
 import React from "react";
 import axios from "axios";
-
-const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users`;
-
-const headers = {
-  headers: {
-    Authorization: "ramon-dias-lovelace",
-  },
-};
+import Cadastro from "./components/CadastroUsuario";
+import Lista from "./components/ListaUsuario";
 
 export default class App extends React.Component {
   state = {
-    id: [],
-    Lnomes: [],
-    Lemail: [],
-    InputNome: "",
-    InputEmail: "",
-  };
-
-  componentDidMount = () => {
-    this.pegarUsuarios();
-  };
-
-  mudaInputNome = (e) => {
-    this.setState({ InputNome: e.target.value });
-  };
-
-  mudaInputEmail = (e) => {
-    this.setState({ InputEmail: e.target.value });
-  };
-
-  pegarUsuarios = () => {
-    axios
-      .get(url, headers)
-      .then((res) => {
-        this.setState({ Lnomes: res.data });
-        console.log(res);
-      })
-      .catch((error) => {
-        alert(error.response.data);
-      });
-  };
-
-  novoUsuario = () => {
-    const body = {
-      name: this.state.InputNome,
-      email: this.state.InputEmail,
-    };
-    axios
-      .post(url, body, headers)
-      .then((res) => {
-        alert("Usuario cadastrado com sucesso!");
-        this.setState({ InputNome: "" });
-        this.setState({ InputEmail: "" });
-        this.pegarUsuarios();
-      })
-      .catch((error) => {
-        alert(error.response.data.message);
-      });
+    home: false
   }
 
-    render() {
-      const componenteCadastro = this.state.Lnomes.map((n) => {
-        return <li key={n.id}>{n.name}</li>;
-      })
+renderizaTela = () => {
+  if (this.state.home) {
+    return <Cadastro cadastro={this.telaCadastro} />;
+  } else {
+    return <Lista listaUsuario={this.telaUsuarios} />;
+  }
+};
 
-      return (
-        <div>
-          <h1>LabeUsers</h1>
-          <input
-            value={this.state.InputNome}
-            onChange={this.mudaInputNome}
-          />
-                    <input
-            value={this.state.InputEmail}
-            onChange={this.mudaInputEmail}
-          />
-          <button onClick={this.novoUsuario}>Enviar</button>
-          {componenteCadastro}
-        </div>
-      )
-    }
+telaCadastro = () => {
+  this.setState({ home: true });
+};
+
+telaUsuarios = () => {
+  this.setState({ home: false });
+};
+
+render() {
+  return (
+    <div>
+      <h1>LabeUsers</h1>
+      <hr />
+      {this.renderizaTela()}
+    </div>
+  );
+}
 }
